CloudLayer uses ShaderBlueprint
-Added layout binding for glsl output (Thanks Jeff 😍)
-CloudLayer now uses a shader blueprint instead of shaderData
